+++ NEWS.md
@@ -1,3 +1,5 @@
+# hatchR (development version)
+
 # hatchR 0.3.2
 
 ## Bug fixes