+++ NEWS.md
@@ -2,7 +2,8 @@
 
 -   adding @pfrater and @allisonswartz as a contributors (should have been done for v0.3.2)
 -   secondary y-axis in `plot_phenology()` now "Cumulative EF Values" and updated documentation (#43)
--   changed name of element `dev.period` to `dev_period` in `predict_phenology()` output; updated name throughout
+-   added new vignette for launching Shiny app (#45)
+-   changed name of element `dev.period` to `dev_period` in `predict_phenology()` output; updated name throughout (#50)
 
 # hatchR 0.3.2