+++ NEWS.md
@@ -1,7 +1,11 @@
 # hatchR (development version)
 
+-   added [Predict spawn](https://bmait101.github.io/hatchR/articles/Predict_spawning.html) vignette for examples of using `predict_spawn()` function
 -   added new function `predict_spawn()` to take observed hatch or spawn events and back calculate when a parent spawned
 -   adding @pfrater and @allisonswartz as a contributors (should have been done for v0.3.2)
+-   secondary y-axis in `plot_phenology()` now "Cumulative EF Values" and updated documentation (#43)
+-   added new vignette for launching Shiny app (#45)
+-   changed name of element `dev.period` to `dev_period` in `predict_phenology()` output; updated name throughout (#50)
 
 # hatchR 0.3.2
 
@@ -14,7 +18,7 @@
 
 ## Bug fixes
 
--   added a missing dependency on R >= 4.1.0 because package code uses the pipe syntax added in R 4.1.0
+-   added a missing dependency on R \>= 4.1.0 because package code uses the pipe syntax added in R 4.1.0
 
 # hatchR 0.3.0